rockchip: rk3568-radxa-e25: Drop PCIe reset-gpios workaround
The GPIO0_C3 pin is changed to use func-3 using pcie30x1m0_pins during
probe of pcie3x1. This cause the device to lock-up when pci driver use
the reset-gpios unless the pin is first changed to use gpio pinmux.

Drop the board u-boot.dtsi workaround now that the gpio and pinctrl
drivers automatically use gpio pinmux when a gpio is requested.

Signed-off-by: Jonas Karlman <jonas@kwiboo.se>