RFC: rockchip: rk3399: Migrate to use TPL on bob and kevin
All Rockchip aarch64 boards, beside bob and kevin, use TPL to initialize
DRAM and SPL to load FIT with U-Boot proper and TF-A.

Bob and kevin currently does not use TPL and instead initialize DRAM
in SPL and directly after tries to load FIT with U-Boot proper and TF-A.

This RFC try to migrate bob and kevin to also use TPL to uniform TPL/SPL
use for all Rockchip aarch64 boards.

Notes and questions:
- Is u-boot.rom even used for bob/kevin? It use u-boot-img, yet SPL
  expect to load FIT and does not permit use of raw imgage.

  Q: Can we remove what looks to be a faulty u-boot.rom definition and
     just use u-boot-rockchip-spi.bin instead?

- FIT location moved from 0x40000 to 0xE0000 in SPI flash to allow for a
  larger SPL, up to 256 KiB, same as on other RK3399 boards.

  Q: Can we use this location?, or is there something external that
    expects fit or u-boot-img at this location?

Features that is now used:
- Use of Rockchip aarch64 common bss and stack addresses
- Checksum/signature check of FIT images in SPL
- Use of dcache to speed up checksum check in SPL
- SDMA use for eMMC access to speed up read/write in U-Boot proper

Changes that would be good for future:
- Include and use pinctrl in SPL
- Use dm gpio instead of spl_gpio functions in SPL

I do not have any bob or kevin so this has only been compile tested.

This should be broken down into smaller logical patches if this is
accepted and works on real hardware.

Signed-off-by: Jonas Karlman <jonas@kwiboo.se>