arm64: dts: rockchip: adjust dcin regulator on ROCK 4D
The ROCK 4D's actual DC input is 5V, and the schematic names it as being
5V as well.

Rename the regulator, and change the voltage it claims to be at.
Furthermore, fix vcc_1v1_nldo_s3's vin-supply as coming from
vcc_5v0_sys, and not the DCIN, as per the schematic. This makes no
functional change; both regulators are always on, and one feeds into the
other.

Signed-off-by: Nicolas Frattaroli <nicolas.frattaroli@collabora.com>
Link: https://lore.kernel.org/r/20250630-rock4d-reg-usb-wifi-v1-1-1057f412d98c@collabora.com
Signed-off-by: Heiko Stuebner <heiko@sntech.de>

[ upstream commit: 9a625a284bfdb902f27f2949063731d189adda3c ]

(cherry picked from commit ac5675c9dfaebe141256577306a0a455e481048a)
Signed-off-by: Jonas Karlman <jonas@kwiboo.se>